bot/pyproject.toml
[tool.poetry]
name = "bot"
authors = ["Interdimensional Walker"]
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
confuse = "^2.0.1"
telethon = "^1.34.0"
izihawa-loglib = "^1.1.4"
izihawa-trident-client = "^1.1.3"
